refactor code with organzied accord
