+++ quora-api/src/main/java/com/upgrad/quora/api/controller/AdminController.java
@@ -1,10 +1,10 @@
 package com.upgrad.quora.api.controller;
 
+import com.upgrad.quora.api.component.AuthorizationHeaderComponent;
 import com.upgrad.quora.api.constants.ResponseMessages;
 import com.upgrad.quora.api.converter.ModelMapperEntityToResponse;
 import com.upgrad.quora.api.model.UserDeleteResponse;
 import com.upgrad.quora.service.business.AdminService;
-import com.upgrad.quora.service.business.AuthorizationHelperService;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AuthorizationFailedException;
 import com.upgrad.quora.service.exception.UserNotFoundException;
@@ -22,14 +22,14 @@ public class AdminController {
     private AdminService adminService;
 
     @Autowired
-    private AuthorizationHelperService authorizationHelperService;
+    private AuthorizationHeaderComponent authorizationHeaderComponent;
 
     @RequestMapping(method = RequestMethod.DELETE,path = "/admin/user/{userId}",
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public ResponseEntity<UserDeleteResponse> userDelete(@PathVariable("userId") final String userUuid,
                                                          @RequestHeader("authorization") final String bearerToken)
             throws AuthorizationFailedException, UserNotFoundException {
-        String accessToken = authorizationHelperService.getBearerToken(bearerToken);
+        String accessToken = authorizationHeaderComponent.getBearerToken(bearerToken);
         UserEntity userEntity = adminService.deleteUser(accessToken,userUuid);
         UserDeleteResponse userDeleteResponse = ModelMapperEntityToResponse.getUserDeleteResponse(
                 userEntity.getUuid(),
