+++ quora-api/src/main/java/com/upgrad/quora/api/controller/CommonController.java
@@ -1,8 +1,8 @@
 package com.upgrad.quora.api.controller;
 
+import com.upgrad.quora.api.component.AuthorizationHeaderComponent;
 import com.upgrad.quora.api.converter.ModelMapperEntityToResponse;
 import com.upgrad.quora.api.model.UserDetailsResponse;
-import com.upgrad.quora.service.business.AuthorizationHelperService;
 import com.upgrad.quora.service.business.CommonService;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AuthorizationFailedException;
@@ -21,13 +21,13 @@ public class CommonController {
     private CommonService commonService;
 
     @Autowired
-    private AuthorizationHelperService authorizationHelperService;
+    private AuthorizationHeaderComponent authorizationHeaderComponent;
 
     @RequestMapping(method = RequestMethod.GET,path = "/userprofile/{userId}",
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public ResponseEntity<UserDetailsResponse> userProfile(@RequestHeader("authorization") final String bearerToken,
                                                            @PathVariable("userId") final String userUuid) throws AuthorizationFailedException, UserNotFoundException {
-        String jwtToken = authorizationHelperService.getBearerToken(bearerToken);
+        String jwtToken = authorizationHeaderComponent.getBearerToken(bearerToken);
         UserEntity userEntity = this.commonService.getUserEntity(jwtToken,userUuid);
         UserDetailsResponse userDetailsResponse = ModelMapperEntityToResponse.getUserDetailsResponse(userEntity);
         return new ResponseEntity<UserDetailsResponse>(userDetailsResponse, HttpStatus.OK);
