+++ quora-api/src/main/java/com/upgrad/quora/api/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.upgrad.quora.api.controller;
 
+import com.upgrad.quora.api.component.AuthorizationHeaderComponent;
 import com.upgrad.quora.api.constants.ResponseMessages;
 import com.upgrad.quora.api.converter.ModelMapperEntityToResponse;
 import com.upgrad.quora.api.converter.ModelMapperRequestToEntity;
@@ -7,7 +8,6 @@ import com.upgrad.quora.api.model.SigninResponse;
 import com.upgrad.quora.api.model.SignoutResponse;
 import com.upgrad.quora.api.model.SignupUserRequest;
 import com.upgrad.quora.api.model.SignupUserResponse;
-import com.upgrad.quora.service.business.AuthorizationHelperService;
 import com.upgrad.quora.service.business.UserBusinessService;
 import com.upgrad.quora.service.entity.UserAuthTokenEntity;
 import com.upgrad.quora.service.entity.UserEntity;
@@ -33,7 +33,7 @@ public class UserController {
     private UserBusinessService userService;
 
     @Autowired
-    private AuthorizationHelperService authorizationHelperService;
+    private AuthorizationHeaderComponent authorizationHeaderComponent;
 
 
     @RequestMapping(method = RequestMethod.POST, path = "/user/signup",
@@ -55,8 +55,8 @@ public class UserController {
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public ResponseEntity<SigninResponse> signin(@RequestHeader("authorization") final String basicToken)
             throws AuthenticationFailedException {
-        final String encodedToken = authorizationHelperService.getBasicToken(basicToken);
-        final String[] userDetails = authorizationHelperService.getUserDetailsFromBearToken(encodedToken);
+        final String encodedToken = authorizationHeaderComponent.getBasicToken(basicToken);
+        final String[] userDetails = authorizationHeaderComponent.getUserDetailsFromBearToken(encodedToken);
         UserAuthTokenEntity userAuthTokenEntity = userService.signin(userDetails[0],userDetails[1]);
         SigninResponse signinResponse = new SigninResponse()
                 .id(userAuthTokenEntity.getUser().getUuid())
@@ -70,7 +70,7 @@ public class UserController {
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public ResponseEntity<SignoutResponse> signout(@RequestHeader("authorization") final  String bearerToken)
             throws SignOutRestrictedException {
-        final String jwtToken = authorizationHelperService.getBearerToken(bearerToken);
+        final String jwtToken = authorizationHeaderComponent.getBearerToken(bearerToken);
         UserAuthTokenEntity userAuthTokenEntity = userService.signout(jwtToken);
         SignoutResponse signoutResponse = new SignoutResponse()
                 .id(userAuthTokenEntity.getUuid())
