+++ quora-service/src/main/java/com/upgrad/quora/service/business/AuthorizationService.java
@@ -1,8 +1,8 @@
 package com.upgrad.quora.service.business;
 
+import com.upgrad.quora.service.component.AuthorizationHelperComponent;
 import com.upgrad.quora.service.constants.ErrorCodeConstants;
 import com.upgrad.quora.service.constants.ErrorMessage;
-import com.upgrad.quora.service.constants.TokenPrefixes;
 import com.upgrad.quora.service.dao.UserAuthTokenDao;
 import com.upgrad.quora.service.dao.UserDao;
 import com.upgrad.quora.service.entity.UserAuthTokenEntity;
@@ -11,11 +11,8 @@ import com.upgrad.quora.service.exception.AuthorizationFailedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.ZonedDateTime;
-import java.util.Base64;
-
 @Service
-public class AuthorizationHelperService {
+public class AuthorizationService {
 
     @Autowired
     private UserAuthTokenDao userAuthTokenDao;
@@ -23,36 +20,13 @@ public class AuthorizationHelperService {
     @Autowired
     private UserDao userDao;
 
+    @Autowired
+    private AuthorizationHelperComponent authorizationHelperComponent;
+
     public UserAuthTokenEntity getUserAuthTokenEntity(final String accesToken) {
         return userAuthTokenDao.getUserAuthToken(accesToken);
     }
 
-    public boolean isValidUserAuthTokenEntity(UserAuthTokenEntity userAuthTokenEntity) {
-        boolean isValid = false;
-        final ZonedDateTime curretTime = ZonedDateTime.now();
-        if ( userAuthTokenEntity != null &&
-                userAuthTokenEntity.getLogoutAt() == null &&
-                userAuthTokenEntity.getExpiresAt().isAfter(curretTime)){
-            isValid = true;
-        }
-        return isValid;
-    }
-
-    public String getBearerToken(final String authorizationHeader){
-        final String encodedJwtToken = authorizationHeader.split(TokenPrefixes.BearerToken.getTokenPrefix())[1];
-        return encodedJwtToken;
-    }
-
-    public String getBasicToken(final String authorizationHeader){
-        final String encodedBasicToken = authorizationHeader.split(TokenPrefixes.BasicToken.getTokenPrefix())[1];
-        return encodedBasicToken;
-    }
-
-    public String[] getUserDetailsFromBearToken(final String basicToken) {
-        final byte[] decodeTokenBytes =  Base64.getDecoder().decode(basicToken);
-        final String[] userDetails = new String(decodeTokenBytes).split(":");
-        return userDetails;
-    }
 
     public UserEntity validateTokenandFetchUserByUUID(final String accessToken,
                                                       final String userUuid) throws AuthorizationFailedException {
@@ -62,7 +36,7 @@ public class AuthorizationHelperService {
                     ErrorMessage.UserHasNotSignedIn.getErrorMessage());
         }
 
-        if(!isValidUserAuthTokenEntity(userAuthTokenEntity)){
+        if(!authorizationHelperComponent.isValidUserAuthTokenEntity(userAuthTokenEntity)) {
             throw new AuthorizationFailedException(ErrorCodeConstants.UserHasSignedOut.getCode(),
                     ErrorMessage.UserHasSignedOut.getErrorMessage());
         }
@@ -70,4 +44,5 @@ public class AuthorizationHelperService {
         UserEntity userEntity = userDao.getUserByUUID(userUuid);
         return  userEntity;
     }
+
 }
