+++ quora-service/src/main/java/com/upgrad/quora/service/component/PasswordCryptographyProvider.java
@@ -1,16 +1,16 @@
-package com.upgrad.quora.service.business;
+package com.upgrad.quora.service.component;
 
 
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.PBEKeySpec;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.spec.InvalidKeySpecException;
 import java.util.Base64;
 import java.util.Random;
-import javax.crypto.SecretKey;
-import javax.crypto.SecretKeyFactory;
-import javax.crypto.spec.PBEKeySpec;
-
-import org.springframework.stereotype.Component;
 
 
 @Component
